+++ openedx/core/djangoapps/edx_global_analytics/utils/utilities.py
@@ -8,11 +8,8 @@ import logging
 from datetime import date, timedelta
 
 import requests
-from django.db.models import Count
-from django.db.models import Q
-from student.models import UserProfile
 
-from openedx.core.djangoapps.edx_global_analytics.utils.cache_utils import cache_instance_data
+from openedx.core.djangoapps.edx_global_analytics.utils.cache_utils import cache_instance_data, get_query_result
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
@@ -23,35 +20,10 @@ def fetch_instance_information(query_type, activity_period, name_to_cache=None):
     Calculate instance information corresponding for particular period as like as previous calendar day and
     statistics type as like as students per country after cached if needed.
     """
-    query_result = get_query_result(query_type, activity_period)
-
-    print(query_result, type(query_result))
-
     if name_to_cache is not None:
-        return cache_instance_data(name_to_cache, query_result)
+        return cache_instance_data(name_to_cache, query_type, activity_period)
 
-    return query_result
-
-
-def get_query_result(query_type, activity_period):
-    """
-    Return query result per query type.
-    """
-    period_start, period_end = activity_period
-
-    if query_type == 'active_students_amount':
-        return UserProfile.objects.exclude(
-            Q(user__last_login=None) | Q(user__is_active=False)
-        ).filter(user__last_login__gte=period_start, user__last_login__lt=period_end).count()
-
-    if query_type == 'students_per_country':
-        return dict(
-            UserProfile.objects.exclude(
-                Q(user__last_login=None) | Q(user__is_active=False)
-            ).filter(user__last_login__gte=period_start, user__last_login__lt=period_end).values(
-                'country'
-            ).annotate(count=Count('country')).values_list('country', 'count')
-        )
+    return get_query_result(query_type, activity_period)
 
 
 def get_previous_day_start_and_end_dates():
